+++ Cargo.toml
@@ -12,7 +12,6 @@ members = [
     "crates/starknet-os",
     "crates/starknet-os-types",
     "tests",
-    "crates/utils",
 ]
 
 [workspace.package]
@@ -82,7 +81,6 @@ starknet-core = "0.11.1"
 starknet-crypto = "0.6.2"
 starknet-os = { path = "crates/starknet-os" }
 starknet-os-types = { path = "crates/starknet-os-types" }
-utils = { path = "crates/utils" }
 starknet-types-core = "0.1.5"
 thiserror = "1.0.48"
 tokio = { version = "1.37.0", features = ["rt-multi-thread"] }
