+++ crates/rpc-replay/Cargo.toml
@@ -14,7 +14,6 @@ serde_json = { workspace = true }
 starknet = { workspace = true }
 starknet_api = { workspace = true }
 starknet-os-types = { workspace = true }
-utils = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 serial_test = "3.2.0"
