--- crates/utils/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "utils"
-version.workspace = true
-edition.workspace = true
-repository.workspace = true
-license.workspace = true
-license-file.workspace = true
-
-[dependencies]
