--- crates/utils/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod env_utils;
